requirements.txt
aiofiles==23.1.0
alembic==1.9.2
anyio==3.6.2
async-timeout==4.0.2
bcrypt==4.0.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Deprecated==1.2.13
dnspython==2.3.0
ecdsa==0.18.0
fastapi-limiter==0.1.5
fastapi==0.89.1
greenlet==2.0.1
h11==0.14.0
idna==3.4
Jinja2==3.1.2
limits==2.8.0
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.2
mongoengine==0.27.0
mysqlclient==2.1.1
packaging==22.0
passlib==1.7.4
pip==23.0.1
pyasn1==0.4.8
pydantic==1.10.4
pymongo==4.3.3
PyMySQL==1.0.2
python-jose==3.3.0
python-multipart==0.0.5
redis==4.5.1
requests==2.28.2
rsa==4.9
setuptools==57.0.0
six==1.16.0
slowapi==0.1.7
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.46
starlette==0.22.0
typing-extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
wheel==0.36.2
wrapt==1.15.0
